renamed aws_hostedZones_client to aws_hosted_zones_client
