+++ tests/test_types_generator.py
@@ -10,7 +10,7 @@ from src.aws_scanner_config import AwsScannerConfig
 from src.clients.aws_iam_client import AwsIamClient
 from src.clients.aws_ec2_client import AwsEC2Client
 from src.clients.aws_logs_client import AwsLogsClient
-from src.clients.aws_hostedZones_client import AwsHostedZonesClient
+from src.clients.aws_hosted_zones_client import AwsHostedZonesClient
 from src.data.aws_athena_data_partition import AwsAthenaDataPartition
 
 from src.data.aws_common_types import Tag
